[fix] : missing required view with id 오류 해결
