+++ core/common/src/main/java/com/hmh/hamyeonham/common/context/ContextExt.kt
@@ -6,9 +6,13 @@ import android.content.pm.PackageManager
 import android.graphics.Point
 import android.graphics.drawable.Drawable
 import android.os.Build
+import android.text.SpannableStringBuilder
+import android.text.Spanned
+import android.text.style.ForegroundColorSpan
 import android.view.View
 import android.view.WindowInsets
 import android.view.WindowManager
+import android.widget.TextView
 import android.widget.Toast
 import androidx.annotation.ColorRes
 import androidx.annotation.DrawableRes
@@ -63,10 +67,9 @@ fun Context.getDeviceSize(): IntArray {
         val windowMetrics = windowManager.currentWindowMetrics
         val windowInsets = windowMetrics.windowInsets
 
-        val insets =
-            windowInsets.getInsetsIgnoringVisibility(
-                WindowInsets.Type.navigationBars() or WindowInsets.Type.displayCutout(),
-            )
+        val insets = windowInsets.getInsetsIgnoringVisibility(
+            WindowInsets.Type.navigationBars() or WindowInsets.Type.displayCutout(),
+        )
         val insetsWidth = insets.right + insets.left
         val insetsHeight = insets.top + insets.bottom
 
@@ -83,13 +86,12 @@ fun Context.getDeviceSize(): IntArray {
     }
 }
 
-fun Context.getAppNameFromPackageName(packageName: String): String =
-    try {
-        val appInfo = packageManager.getApplicationInfo(packageName, PackageManager.GET_META_DATA)
-        packageManager.getApplicationLabel(appInfo).toString()
-    } catch (e: PackageManager.NameNotFoundException) {
-        "Unknown"
-    }
+fun Context.getAppNameFromPackageName(packageName: String): String = try {
+    val appInfo = packageManager.getApplicationInfo(packageName, PackageManager.GET_META_DATA)
+    packageManager.getApplicationLabel(appInfo).toString()
+} catch (e: PackageManager.NameNotFoundException) {
+    "Unknown"
+}
 
 fun Context.getAppIconFromPackageName(packageName: String): Drawable? {
     try {
@@ -99,4 +101,22 @@ fun Context.getAppIconFromPackageName(packageName: String): Drawable? {
         e.printStackTrace()
     }
     return ContextCompat.getDrawable(this, R.drawable.ic_launcher_foreground)
+}
+
+fun Context.concatDifferentColored남음AndBindText(str: String, tv: TextView, color: Int) {
+    val builder = SpannableStringBuilder(
+        str,
+    )
+    builder.setSpan(
+        ForegroundColorSpan(
+            ContextCompat.getColor(
+                this,
+                color,
+            ),
+        ),
+        builder.length - 2,
+        builder.length,
+        Spanned.SPAN_INCLUSIVE_EXCLUSIVE,
+    )
+    tv.text = builder
 }
