+++ feature/main/src/main/java/com/hmh/hamyeonham/feature/main/home/UsageStaticsTotalViewHolder.kt
@@ -1,15 +1,10 @@
 package com.hmh.hamyeonham.feature.main.home
 
 import android.content.Context
-import android.text.SpannableStringBuilder
-import android.text.Spanned.SPAN_INCLUSIVE_EXCLUSIVE
-import android.text.style.ForegroundColorSpan
-import androidx.core.content.ContextCompat.getColor
-import androidx.core.content.ContextCompat.getString
 import androidx.recyclerview.widget.RecyclerView
+import com.hmh.hamyeonham.common.context.concatDifferentColored남음AndBindText
 import com.hmh.hamyeonham.common.time.convertTimeToString
 import com.hmh.hamyeonham.common.view.initAndStartProgressBarAnimation
-import com.hmh.hamyeonham.feature.main.R
 import com.hmh.hamyeonham.feature.main.databinding.ItemUsagestaticTotalBinding
 import com.hmh.hamyeonham.usagestats.model.UsageStatAndGoal
 
@@ -19,34 +14,17 @@ class UsageStaticsTotalViewHolder(
 ) : RecyclerView.ViewHolder(binding.root) {
     fun onBind(usageStatAndGoal: UsageStatAndGoal) {
         binding.run {
-            tvItemusagestaticTotalHour.text = convertTimeToString(usageStatAndGoal.goalTime)
-            pbItemusagestaticTotal.progress = usageStatAndGoal.usedPercentage
+            tvTotalGoal.text = convertTimeToString(usageStatAndGoal.goalTime)
+            pbTotalUsage.progress = usageStatAndGoal.usedPercentage
         }
-        bindTotalLeftWithSpannable(usageStatAndGoal)
+        context.concatDifferentColored남음AndBindText(
+            convertTimeToString(usageStatAndGoal.goalTime),
+            binding.tvTotalTimeLeft,
+            com.hmh.hamyeonham.core.designsystem.R.color.gray1,
+        )
         initAndStartProgressBarAnimation(
-            binding.pbItemusagestaticTotal,
+            binding.pbTotalUsage,
             usageStatAndGoal.usedPercentage,
         )
-    }
-
-    fun bindTotalLeftWithSpannable(usageStatAndGoal: UsageStatAndGoal) {
-        val builder = SpannableStringBuilder(
-            convertTimeToString(usageStatAndGoal.timeLeft) + " " + getString(
-                context,
-                R.string.all_left,
-            ),
-        )
-        builder.setSpan(
-            ForegroundColorSpan(
-                getColor(
-                    context,
-                    com.hmh.hamyeonham.core.designsystem.R.color.gray1,
-                ),
-            ),
-            builder.length - 2,
-            builder.length,
-            SPAN_INCLUSIVE_EXCLUSIVE,
-        )
-        binding.tvItemusagestaticTotalTimeLeft.text = builder
     }
 }
