+++ feature/main/src/main/java/com/hmh/hamyeonham/feature/main/home/UsageStaticsViewHolder.kt
@@ -1,16 +1,12 @@
 package com.hmh.hamyeonham.feature.main.home
 
 import android.content.Context
-import android.text.SpannableStringBuilder
-import android.text.Spanned
-import android.text.style.ForegroundColorSpan
-import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
+import com.hmh.hamyeonham.common.context.concatDifferentColored남음AndBindText
 import com.hmh.hamyeonham.common.context.getAppIconFromPackageName
 import com.hmh.hamyeonham.common.context.getAppNameFromPackageName
 import com.hmh.hamyeonham.common.time.convertTimeToString
 import com.hmh.hamyeonham.common.view.initAndStartProgressBarAnimation
-import com.hmh.hamyeonham.feature.main.R
 import com.hmh.hamyeonham.feature.main.databinding.ItemUsagestaticBinding
 import com.hmh.hamyeonham.usagestats.model.UsageStatAndGoal
 
@@ -20,38 +16,20 @@ class UsageStaticsViewHolder(
 ) : RecyclerView.ViewHolder(binding.root) {
     fun onBind(usageStatAndGoal: UsageStatAndGoal) {
         binding.run {
-            tvItemusagestatAppname.text =
-                context.getAppNameFromPackageName(usageStatAndGoal.packageName)
-            ivItemusagestatAppicon.setImageDrawable(
+            tvAppname.text = context.getAppNameFromPackageName(usageStatAndGoal.packageName)
+            ivAppicon.setImageDrawable(
                 context.getAppIconFromPackageName(
                     usageStatAndGoal.packageName,
                 ),
             )
-            tvItemusagestatGoalTime.text = convertTimeToString(usageStatAndGoal.goalTime)
-            pbItemUsagestat.progress = usageStatAndGoal.usedPercentage
+            tvAppGoalTime.text = convertTimeToString(usageStatAndGoal.goalTime)
+            pbAppUsage.progress = usageStatAndGoal.usedPercentage
         }
-        bindTotalLeftWithSpannable(usageStatAndGoal)
-        initAndStartProgressBarAnimation(binding.pbItemUsagestat, usageStatAndGoal.usedPercentage)
-    }
-
-    fun bindTotalLeftWithSpannable(usageStatAndGoal: UsageStatAndGoal) {
-        val builder = SpannableStringBuilder(
-            convertTimeToString(usageStatAndGoal.timeLeft) + " " + ContextCompat.getString(
-                context,
-                R.string.all_left,
-            ),
+        context.concatDifferentColored남음AndBindText(
+            convertTimeToString(usageStatAndGoal.goalTime),
+            binding.tvAppTimeLeft,
+            com.hmh.hamyeonham.core.designsystem.R.color.gray1,
         )
-        builder.setSpan(
-            ForegroundColorSpan(
-                ContextCompat.getColor(
-                    context,
-                    com.hmh.hamyeonham.core.designsystem.R.color.gray2,
-                ),
-            ),
-            builder.length - 2,
-            builder.length,
-            Spanned.SPAN_INCLUSIVE_EXCLUSIVE,
-        )
-        binding.tvItemusagestatTimeLeft.text = builder
+        initAndStartProgressBarAnimation(binding.pbAppUsage, usageStatAndGoal.usedPercentage)
     }
 }
