+++ feature/statistics/src/main/java/com/hmh/hamyeonham/statistics/StaticsActivity.kt
@@ -3,13 +3,10 @@ package com.hmh.hamyeonham.statistics
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.lifecycleScope
-import androidx.recyclerview.widget.LinearLayoutManager
 import com.hmh.hamyeonham.common.navigation.NavigationProvider
 import com.hmh.hamyeonham.common.view.viewBinding
 import com.hmh.hamyeonham.feature.statistics.databinding.ActivityStaticsBinding
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -23,19 +20,5 @@ class StaticsActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
-        collectUsageStatsList()
-    }
-
-    private fun collectUsageStatsList() {
-        val usageStaticsAdapter = UsageStaticsAdapter()
-        binding.rvStatics.run {
-            adapter = usageStaticsAdapter
-            layoutManager = LinearLayoutManager(this@StaticsActivity)
-        }
-        lifecycleScope.launch {
-            staticsViewModel.usageStatAndGoalList.collect {
-                usageStaticsAdapter.submitList(it)
-            }
-        }
     }
 }
