--- feature/statistics/src/main/res/layout/item_usagestatic.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginHorizontal="20dp">
-
-    <com.google.android.material.progressindicator.LinearProgressIndicator
-        android:id="@+id/pb_item_usagestat"
-        style="@style/Widget.Hmh.ProgressIndicator"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="12dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
-        android:id="@+id/iv_itemusagestat_appicon"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_marginStart="17dp"
-        android:layout_marginTop="17dp"
-        android:layout_marginBottom="16dp"
-        android:contentDescription="@string/item_usagestatic_appicon_description"
-        app:layout_constraintBottom_toBottomOf="@id/pb_item_usagestat"
-        app:layout_constraintDimensionRatio="1"
-        app:layout_constraintStart_toStartOf="@id/pb_item_usagestat"
-        app:layout_constraintTop_toTopOf="@id/pb_item_usagestat" />
-
-    <TextView
-        android:id="@+id/tv_itemusagestat_appname"
-        style="?textAppearanceLabelLarge"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="9dp"
-        android:layout_marginTop="2dp"
-        android:textColor="?gray1"
-        app:layout_constraintStart_toEndOf="@id/iv_itemusagestat_appicon"
-        app:layout_constraintTop_toTopOf="@id/iv_itemusagestat_appicon" />
-
-    <TextView
-        android:id="@+id/tv_itemusagestat_goal_time"
-        style="?textAppearanceBodySmall"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="2dp"
-        android:textColor="?white_text"
-        app:layout_constraintBottom_toBottomOf="@id/iv_itemusagestat_appicon"
-        app:layout_constraintStart_toStartOf="@id/tv_itemusagestat_appname" />
-
-    <TextView
-        android:id="@+id/tv_itemusagestat_time_left"
-        style="?textAppearanceTitleMedium"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="20dp"
-        android:textColor="?white_text"
-        app:layout_constraintBottom_toBottomOf="@id/pb_item_usagestat"
-        app:layout_constraintEnd_toEndOf="@id/pb_item_usagestat"
-        app:layout_constraintTop_toTopOf="@id/pb_item_usagestat" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
