--- feature/statistics/src/main/res/layout/item_usagestatic_total.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <ImageView
-        android:id="@+id/iv_itemusagestatic_total_blackhole"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:contentDescription="@string/item_usagestatic_total_blackhole_description"
-        android:src="@drawable/img_statics_blackhole"
-        app:layout_constraintDimensionRatio="1"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/tv_itemusagestatic_total_blackhole"
-        style="?textAppearanceDisplaySmall"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="33dp"
-        android:text="@string/item_usagestatic_total_blackhole"
-        android:textColor="?white_text"
-        app:layout_constraintStart_toStartOf="@id/tv_itemusagestatic_total_goal_title"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <com.google.android.material.progressindicator.LinearProgressIndicator
-        android:id="@+id/pb_itemusagestatic_total"
-        style="@style/Widget.Hmh.TotalProgressIndicator"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="20dp"
-        android:layout_marginTop="27dp"
-        android:layout_marginBottom="33dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/iv_itemusagestatic_total_blackhole" />
-
-    <TextView
-        android:id="@+id/tv_itemusagestatic_total_hour"
-        style="?textAppearanceDisplayMedium"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="10dp"
-        android:textColor="?white_text"
-        app:layout_constraintBottom_toTopOf="@id/pb_itemusagestatic_total"
-        app:layout_constraintStart_toStartOf="@id/pb_itemusagestatic_total" />
-
-    <TextView
-        android:id="@+id/tv_itemusagestatic_total_goal_title"
-        style="?textAppearanceLabelMedium"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/all_goal_time"
-        android:textColor="?gray1"
-        app:layout_constraintBottom_toTopOf="@id/tv_itemusagestatic_total_hour"
-        app:layout_constraintStart_toStartOf="@id/pb_itemusagestatic_total" />
-
-    <TextView
-        android:id="@+id/tv_itemusagestatic_total_time_left"
-        style="?textAppearanceBodySmall"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="12dp"
-        android:textColor="?white_text"
-        app:layout_constraintBottom_toTopOf="@id/pb_itemusagestatic_total"
-        app:layout_constraintEnd_toEndOf="@id/pb_itemusagestatic_total" />
-</androidx.constraintlayout.widget.ConstraintLayout>
